+++ module/data/actor-npc.mjs
@@ -96,6 +96,7 @@ export default class SdmNPC extends SdmActorBase {
     data.defense = this.defense;
     data.level = this.level;
     data.life = this.life;
+    data.system = this;
 
     return data;
   }