+++ module/data/base-actor.mjs
@@ -1,3 +1,5 @@
+import { SizeUnit } from '../helpers/constants.mjs';
+
 export default class SdmActorBase extends foundry.abstract.TypeDataModel {
   static LOCALIZATION_PREFIXES = ['SDM.Actor.base'];
 
@@ -7,6 +9,18 @@ export default class SdmActorBase extends foundry.abstract.TypeDataModel {
 
     schema.biography = new fields.HTMLField();
     schema.notes = new fields.HTMLField();
+
+    schema.size = new fields.SchemaField({
+      value: new foundry.data.fields.NumberField({ required: true, initial: 1, min: 0 }),
+      unit: new foundry.data.fields.StringField({
+        required: true,
+        initial: SizeUnit.SACKS,
+        choices: Object.values(SizeUnit).reduce((acc, key) => {
+          acc[key] = game.i18n.localize(CONFIG.SDM.sizeUnits[key]);
+          return acc;
+        }, {})
+      })
+    });
 
     return schema;
   }